pyproject.toml
[tool.poetry]
name = "pygemini"
version = "0.1.0"
description = ""
authors = ["Shuichi Kataoka <shu01k9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-generativeai = "^0.5.2"
python-dotenv = "^1.0.1"
gradio = "^4.28.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
